Add test information table with concatenated IDs
This test information table has a column with ID1 and ID2, or ID1, ID2 and ID3 concatenated. All spaces in this column are ("manually") found and replaced with "_", i.e. a single underscore. The columns are concatenated with a "__" separator, i.e. two underscores. The column that contains the arguments of the composite score was updated to reflect this. For now, this template can coexist with the template that is already there in this repository, as that one might still be useful in testing other features.