Check whether email address in signup form is on the whitelist
Added a check that tests whether the email address used in the signup
form is on the email adress whitelist.

To create an account the email domain must be on the domain whitelist
or the address must be on the email address whitelist. Otherwise
an appropriate error message is displayed. This message still needs to
be adjusted.