Add explanation to legend
A line explaining that patients can be clicked was added to the
legend. This also fixes the problem that the area for the first
patient that can be clicked was really small (only the last part
of the patient id could be clicked to select/deselect that patient).
It was not clear why this problem occurred.