Disable/enable input fields for intermediary or computed values
Input fields for intermediary or computed valued are enabled or
disabled based on whether the user fills out a value. If the user
fills out an intermediary value, the computed value input is
disabled. If he fills out the computed value, the intermediary
values are disabled. Disabled computed value field is enabled if
all intermediary value fields are empty and the intermediary value
fields are enabled if the computed value field is empty.