Fix bug in data upload service
Due to a bug, the number of (empty) patients that were intitally added
to the patients array was incorrect. This caused the data upload to fail.
The error was fixed by adding an empty patient if the length of the
patients array is smaller than or equal to the index of the current
patient.