Trigger calculation of computed variables when user clicks spinner
The enabling and disabling of input fields for intermediary/computed
values is also triggered when the number up and down arrows (spinners)
are clicked.

Because spinners are added by the browser, and disabling spinners is
browser dependent, we decided to fix this problem by also triggering
validation when the arrows are clicked.

Fixes #186
See also #184