Make ellipse parameters depend on normative data version and conf
The parameters used to draw the ellipses did not depend on normative
data version or the confidence interval. To solve this, ellipse parameters
for the 2016 version and different confidence intervals was added to the
data directory. The code was adapted to load these files.

Because the tests for normative data versions 2015 and 2014 do not have all
the properties that are available for 2016, some additional checking is done.

Refs #167 #168

What needs to be done:
- add ellipses parameters files for different confidence intervals for
the other normative data versions
- fix error with the ocpu service and normative data versions 2015 and 2014