Fix disabling of input fields when user uses tab to navigate
Both input fields for intermediary values and the computed value were
disabled if the user uses tab to navigate from the last intermediary value
to the (disabled/readonly) computed value. This was fixed by checking
that the computed value is not already disabled before disabling the
input fields of the intermediary values.

Refs the reopened #184